requirements.txt
streamlit
pyngrok
google-generativeai
python-dotenv

// Run the following command to install the required packages
// pip install -r requirements.txt
